requirements.txt
discord.py==1.2.3
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.10.3
